Refactored content panel for smaller components, changed surfaces to have right color
