+++ frontend/src/components/Surface/index.tsx
@@ -21,7 +21,7 @@ export const Surface: React.FC<SurfaceProps> = ({
     className,
 }) => {
     return (
-        <div className={`bg-primary-95 ` + className}>
+        <div className={`bg-primary-99 ` + className}>
             <div
                 className={classNames(
                     'bg-primary',
